Rename the _coconext extension module to _pycoconext
The nanobind extension was named `_coconext`, one character away from
the `libcoconext` C++ library it wraps. Once libcoconext becomes a
shared object sitting next to the extension in site-packages, two files
differing only by a `lib` prefix and a `py` infix is a recipe for
confusion (and for the wrong one being preloaded/linked).

Rename the extension to `_pycoconext` to make the Python-binding layer
unambiguous against the pure-C++ `libcoconext`. Pure rename: the module
contents, stub, and public `coconext.types` API are unchanged.
